Move `AbstractRobot` and `Layout` to shared source set
